The Unarchiver to expand all the things 🚚
The Unarchiver is a much more capable replacement for "Archive
Utility.app", the built-in archive unpacker program on Mac OS X. The
Unarchiver is designed to handle many more formats than Archive Utility,
and to better fit in with the design of the Finder. It can also handle
filenames in foreign character sets, created with non-English versions
of other operating systems. I personally find it useful for opening
Japanese archives, but it should handle many other languages just as
well.

http://wakaba.c3.cx/s/apps/unarchiver